Add Action<MidiMessage> Change namespace, Dono.Midi.Types => Dono.Midi

+++ Runtime/Utilities/MidiUtilities.MessageEditor.cs
@@ -2,8 +2,9 @@ namespace Dono.Midi
 {
     public static partial class MidiUtilities
     {
-        
-        public static void Transpose(out MidiMessage message, byte value)
-        { throw new System.NotImplementedException(); }
+        public static void Transpose(out MidiMessage message, int value)
+        {
+            throw new System.NotImplementedException();
+        }
     }
 }
